==> cross_tol_data/9_predictions.Rmd <==
---
title: "predictions"
output: html_document
date: '2023-06-06'
---
```{r setup, include=FALSE, echo = FALSE}
#comment
knitr::opts_chunk$set(echo = TRUE)
library(ggplot2)
library(viridis)
library(cowplot)
library(tidyverse)
library(vegan)
library(ggrepel)
library(dplyr)
library(gtools)
library(reshape)
library(reshape2)
library(lme4)
library(lmerTest)
library(tidyr)
library(broom)
library(rsq)
library(performance)
library(Rmisc)


x = getwd()
setwd(x)
```

Setting color palette
```{r}
relfitPalette <- c("#787878", "#FF9A17", "#3B83FF","#3EFF3B", "#FD73FF", "#FFF200", "#A880FF")
relfitPalette2 <- c("#FF9A17", "#3B83FF","#3EFF3B", "#FD73FF", "#FFF200", "#A880FF")

legend_title1 <- "Test\nEnvironment"
legend_title2 <- "Evolution\nEnvironment"
legend_title3 = "Environment"
```

get those names right
```{r }

name_codes = read_csv("code_strain_line.csv")
mydata3 <- read.csv("mbm_intermediate_files/mydata3.csv")

```


adding SGD prediction data and usgs correlations petite info for model (different than visualization)
```{r}
prop_pleio <- read.csv("predictions/prop_pleio_both.csv")
prop_pleiovis <- read.csv("predictions/prop_pleio.csv")
usgs = read_csv('predictions/soilmet_trans_combos_both.csv')
usgsvis = read_csv('predictions/soilmet_trans_combos.csv')
ORP = read_csv('predictions/elect_pot_both.csv')
ORP$diff = -(ORP$sim)
ORPvis = read_csv('predictions/ele_pot.csv')
ORPvis$diff = -(ORPvis$ORP)

met_ppm = read_csv("predictions/met_ppm.csv")

petite = read_csv('raw_outcross_petite/petites.csv')
colnames(petite)[colnames(petite) == "evoline"] ="code_0"
petites <- merge(petite,name_codes,by="code_0")
colnames(petites)[colnames(petites) == "code_2"] ="strain"

```

# data to calculate cross-tolerance for each evolved line


```{r}


mydata3$combo <- paste(mydata3$E1, mydata3$E2, sep= "")
mydata3$prop.syn <- prop_pleio$prop.syn[match(mydata3$combo,prop_pleio$combo)] 
mydata3$no.genes <- prop_pleio$no.genes[match(mydata3$combo,prop_pleio$combo)] 
mydata3$all_slope <- usgs$all_slope[match(mydata3$combo,usgs$combo)] 
mydata3$ORP <- ORP$diff[match(mydata3$combo,ORP$combo)] 
mydata3$petite <- petites$petite[match(mydata3$strain,petites$strain)] 

react.norm =mydata3
mydata3=na.omit(mydata3)


var_data3 <- mydata3 %>%
  filter(E2!=".YPAD") 

```


# look at the data

```{r}
#data3 = read_csv("data3.csv")

#just look at the data, 
hist(var_data3$relfit) # looks super normal
hist(log(var_data3$relfit)) # that's better

hist(var_data3$ORP) # this looks bimodal and I don't know
hist(var_data3$prop.syn) # this almost looks normal?
hist(var_data3$all_slope) # 
```


plot data against relfit1

```{r}
plot(var_data3$ORP,var_data3$relfit) #is this a negative slope???? so the more different the oxidative reduction potential of the media are, the more cross-tolerance?? UGH i don't know
plot(var_data3$prop.syn,var_data3$relfit)
plot(var_data3$all_slope,var_data3$relfit)

var_lm = lm(log(relfit+1) ~ E2+E1+E1:E2+machine,data=var_data3)
summary(var_lm)

```


```{r}
#test random effects on crosstolerance include measurement error


var_lmer <- lmer(1 + log(relfit+1) ~ ORP + prop.syn + all_slope  + petite + (1|E2)+(1|E1)+(1|E1:E2)+(1|machine/plate)+(1|strain:E2)+(1|strain:E1)+(1|strain:E1:E2),data=var_data3)
summary(var_lmer)
icc(var_lmer, by_group = TRUE)
anova(var_lmer)

```


#Cross tolerance new calculation

```{r}
cross.tol.combos = read_csv("mbm_intermediate_files/pairs_cross_tolerance.csv")
usgsvis$comba <- paste(usgsvis$reference, usgsvis$combo, sep= "")


cross.tol.combos$prop.syn <- prop_pleiovis$prop.syn[match(cross.tol.combos$combo,prop_pleiovis$combo)] 
cross.tol.combos$no.genes <- prop_pleiovis$no.genes[match(cross.tol.combos$combo,prop_pleiovis$combo)] 
cross.tol.combos$all_slope <- usgsvis$all_slope[match(cross.tol.combos$combo,usgsvis$comba)] 
cross.tol.combos$ORP <- ORPvis$diff[match(cross.tol.combos$combo,ORPvis$combo)] 

lm_ct =lm(ct ~ prop.syn + all_slope + ORP, data=cross.tol.combos)
lm_ct_prop =lm(ct ~ prop.syn , data=cross.tol.combos)
lm_ct_slope =lm(ct ~ all_slope , data=cross.tol.combos)
lm_ct_orp =lm(ct ~  ORP, data=cross.tol.combos)
summary(lm_ct)
summary(lm_ct_prop)
summary(lm_ct_slope)
summary(lm_ct_orp)
```

# plot the linear models of the pairs of metals against the predictions
```{r}

ct_orp =cross.tol.combos %>% 
  ggplot(aes(x=ORP,y=ct)) +
    geom_point() +
    geom_smooth(method="lm",linetype="dashed",color="black", se=FALSE) +
     theme_minimal() +
    labs(x="Electropotential similarity", y="Metal pairs\nCross-tolerance")+
  theme(axis.text=element_text(size=16),
        axis.title=element_text(size=18))
ct_orp

ct_slope =cross.tol.combos %>% 
  ggplot(aes(x=all_slope,y=ct)) +
    geom_point() +
    geom_smooth(method="lm",linetype="dashed",color="black", se=FALSE) +
     theme_minimal() +
    labs(x="Metal co-occurrence", y="Metal pairs\nCross-tolerance")+
  theme(axis.text=element_text(size=16),
        axis.title=element_text(size=18))
ct_slope

ct_prop =cross.tol.combos %>% 
  ggplot(aes(x=prop.syn,y=ct)) +
    geom_point() +
    geom_smooth(method="lm",linetype="dashed",color="black", se=FALSE) +
     theme_minimal() +
    labs(x="Proportion of gene overlap", y="Metal pairs\nCross-tolerance")+
  theme(axis.text=element_text(size=16),
        axis.title=element_text(size=18))
ct_prop


```


Data visualization
```{r}

orplin=mydata3 %>% 
  ggplot(aes(x=ORP,y=relfit)) +
    geom_point() +
    geom_smooth(method="lm",linetype="dashed",color="black", se=FALSE) +
     theme_minimal() +
  theme(axis.text=element_text(size=14),
        axis.title=element_text(size=14,face="bold"))+
    labs(x="Electropotential similarity", y="Cross-tolerance per evolved line")


proplin=mydata3 %>% 
  ggplot(aes(x=prop.syn,y=relfit)) +
    geom_point() +
    geom_smooth(method="lm",linetype="dashed",color="black", se=FALSE) +
     theme_minimal() +
  theme(axis.text=element_text(size=14),
        axis.title=element_text(size=14,face="bold"))+
    labs(x="Proportion of gene overlap", y="Cross-tolerance per evolved line")


usgslin=mydata3 %>% 
  ggplot(aes(x=all_slope,y=relfit)) +
    geom_point() +
    geom_smooth(method="lm",linetype="dashed",color="black", se=FALSE) +
     theme_minimal() +
  theme(axis.text=element_text(size=14),
        axis.title=element_text(size=20))+
    labs(x="Metal co-occurrence", y="Cross-tolerance per evolved line")




```


```{r}
# Make heatmap of the differences between metal reactivity values

orpmap=ORPvis %>% 
  ggplot(aes(x=met1,y=met2,color=diff)) + 
    geom_point(shape=15, size=15) +
    scale_color_gradient(low = "yellow", high = "blue",
    name="Electropotential\nsimilarity") +
    scale_x_discrete(position = "top") + 
    theme_bw() +
    theme(axis.title.x=element_blank(), axis.title.y=element_blank()) +
  theme(axis.text=element_text(size=20),
        axis.title=element_text(size=22))
    #+scale_size(range=c(8,18), name="Number of genes\nin common")
#ggsave("~/Dropbox/ottoLabNotes/aneuploidy/ORPheatmap.jpeg", width=8, height=6, units = "in", dpi=1000)

propmap=prop_pleiovis %>% 
  ggplot(aes(x=met1,y=met2,color=prop.syn)) + 
    geom_point(aes(size = no.genes), shape=15) +
    scale_color_gradient(low = "yellow", high = "blue",
    name="Proportion\nof gene overlap") +
    scale_x_discrete(position = "top") + 
    theme_bw() +
    theme(axis.title.x=element_blank(), axis.title.y=element_blank())+
  theme(axis.text=element_text(size=20),
        axis.title=element_text(size=22)) +
    scale_size(range=c(8,18), name="Number of genes\nin common")


usgsmap=usgsvis %>% 
  ggplot(aes(x=reference,y=combo,color=all_slope)) + 
    geom_point(shape=15, size=15) +
    scale_color_gradient(low = "yellow", high = "blue",
    name="Metal\nco-occurrence") +
    scale_x_discrete(position = "top") + 
    theme_bw() +
    theme(axis.title.x=element_blank(), axis.title.y=element_blank()) +
  theme(axis.text=element_text(size=20),
        axis.title=element_text(size=22))


```



```{r }
p1=plot_grid(
  orpmap, ct_orp, labels = c("A","B"),
  ncol = 2
)
p2=plot_grid(propmap,ct_prop,  labels = c("C","D"),
  ncol = 2
)
p3=plot_grid(usgsmap,ct_slope, labels = c("E","F"),
  ncol = 2
)

plot_grid(p1, p2, p3, ncol = 1)

ggsave("figures/prediction_grid.jpeg", width=15, height=13, units = "in", dpi=700)
```
